Add files via upload
This scripts were written by the MSLIS student we briefly had on the project. They're designed to take a mention, query the NCBI taxonomy DB's website, spider the result's page for a match and then scrape the taxID from the subsequent page. They never worked well (partially because the NCBI pre-canned queries are sub-optimal). This code was all deprecated in favor of tried and true manual cleaning methods.